crates/minimald: create scaffolding for building up SSH channels

+++ crates/minimald/src/main.rs
@@ -7,10 +7,7 @@ use std::path::PathBuf;
 use tokio::net::UnixListener;
 use tracing_subscriber::{EnvFilter, fmt, prelude::*};
 
-use crate::server::Server;
-
-mod connection;
-mod server;
+use minimald::server::{Config, HostKey, Server};
 
 #[derive(Parser)]
 #[command(name = "minimald", version = env!("CARGO_PKG_VERSION"), long_version = env!("LONG_VERSION"))]
@@ -26,19 +23,36 @@ struct Cli {
 impl Cli {
     /// Returns the path to the minimal-dir (base directory for state)
     /// based on command-line arguments.
-    pub fn minimal_dir(&self) -> PathBuf {
+    pub fn minimal_state_dir(&self) -> PathBuf {
         self.global_args.minimal_dir.clone().unwrap_or_else(|| {
-            dirs::cache_dir()
-                .unwrap_or_else(|| PathBuf::from("~/.cache"))
+            dirs::state_dir()
+                .unwrap_or_else(|| PathBuf::from("~/.local/state"))
                 .join("minimal")
         })
     }
+
+    /// Returns the path to base directory for caching
+    /// based on command-line arguments.
+    pub fn minimal_cache_dir(&self) -> PathBuf {
+        dirs::cache_dir()
+            .unwrap_or_else(|| PathBuf::from("~/.local/cache"))
+            .join("minimal")
+    }
+
+    /// Returns the path to the directory containing sockets/info about this daemon for clients.
+    pub fn client_instance_dir(&self) -> PathBuf {
+        let instance_num = match &self.command {
+            Command::Run(ListenArgs { instance_num }) => *instance_num,
+            _ => 0,
+        };
+        self.minimal_state_dir()
+            .join("providers")
+            .join(format!("local-{instance_num}"))
+    }
+
     /// Returns the path to the UDS socket we should listen on.
     pub fn listen_on(&self) -> PathBuf {
-        match &self.command {
-            Command::Run(ListenArgs { socket: Some(p) }) => p.clone(),
-            _ => self.minimal_dir().join("minimald.sock"),
-        }
+        self.client_instance_dir().join("ssh.sock")
     }
 }
 
@@ -64,7 +78,7 @@ struct CompletionsArgs {
 /// Shared arguments for all subcommands.
 #[derive(Debug, Args)]
 pub struct GlobalArgs {
-    /// Override the base directory used for operations (default: ~/.cache/minimal)
+    /// Override the state directory used for operations (default: $XDG_STATE_DIR/minimal)
     #[arg(long)]
     minimal_dir: Option<PathBuf>,
     /// Load the minimal standard library from the given path instead
@@ -80,9 +94,12 @@ pub struct GlobalArgs {
 /// Arguments describing where minimald should listen for connections.
 #[derive(Debug, Args)]
 pub struct ListenArgs {
-    /// Override the UDS path to listen on.
-    #[arg(long)]
-    socket: Option<PathBuf>,
+    /// Instance number for this minimald; determines client-relevant paths under
+    /// `<minimal_state_dir>/providers/local-<instance-num>`.
+    ///
+    /// The SSH socket is accessible as `ssh.sock`.
+    #[arg(long, default_value_t = 0)]
+    instance_num: u32,
 }
 
 /// An error at the top level of minimald.
@@ -114,7 +131,7 @@ async fn main() -> Result<(), MainError> {
         return Ok(());
     }
 
-    if let Err(e) = std::fs::create_dir_all(cli.minimal_dir())
+    if let Err(e) = std::fs::create_dir_all(cli.minimal_state_dir())
         && e.kind() != std::io::ErrorKind::AlreadyExists
     {
         return Err(MainError::IO(e, "creating minimal dir"));
@@ -122,6 +139,15 @@ async fn main() -> Result<(), MainError> {
 
     // If we got this far we need to launch minimald.
     //
+    // Ensure the socket's parent directory exists.
+    let socket_path = cli.listen_on();
+    if let Some(parent) = socket_path.parent()
+        && let Err(e) = std::fs::create_dir_all(parent)
+        && e.kind() != std::io::ErrorKind::AlreadyExists
+    {
+        return Err(MainError::IO(e, "creating provider dir"));
+    }
+
     // Listen on the UDS socket.
     if let Err(e) = std::fs::remove_file(cli.listen_on())
         && e.kind() != std::io::ErrorKind::NotFound
@@ -131,11 +157,21 @@ async fn main() -> Result<(), MainError> {
     let listener =
         UnixListener::bind(cli.listen_on()).map_err(|e| MainError::IO(e, "listening to socket"))?;
     tracing::info!("Started listening on {}", cli.listen_on().display());
+    tracing::info!(
+        "Run the following to debug the socket:\n\nssh -o ProxyCommand='socat - UNIX-CONNECT:{}' \\\n\t-o 'StrictHostKeyChecking=no' -o 'UserKnownHostsFile=/dev/null' \\\n\tlocal",
+        cli.listen_on().display()
+    );
+
+    // TODO: When we have a daemonize command, daemonize here.
 
     // Setup the server.
-    let config = server::Config {
-        host_key: server::HostKey::Ephemeral,
-        minimal_dir: cli.minimal_dir(),
+    let config = Config {
+        host_key: HostKey::OnDisk {
+            path: cli.client_instance_dir().join("ssh_host_ed25519_key"),
+            create_if_missing: true,
+        },
+        minimal_state_dir: cli.minimal_state_dir(),
+        minimal_cache_dir: cli.minimal_cache_dir(),
     };
 
     match cli.command {
